All the int's are now int16_t Added the == operator in simple cards fixed all the errors and added in the main file the necesary data in order to run training.

+++ Eter/Player.cpp
@@ -207,7 +207,7 @@ SimpleCard Player::chooseCard()
 
 int Player::numberofValidCards()
 {
-	int count = 0;
+	int16_t count = 0;
 	for (auto& card : m_simpleCardsVector)
 	{
 		if (card.getColor() == Color::Red || card.getColor() == Color::Blue)
